+++ libs/gesha-core/src/gateway/mod.rs
@@ -9,3 +9,6 @@ pub use writer::Writer;
 
 mod diff;
 pub use diff::detect_diff;
+
+mod tests;
+pub use tests::{run_test, TestTarget};
